Limit results to tags
There is now a form on every panel that can be toggled with the "edit" button. It brings up a text box. Enter tags to include, separated by a comma and space.

New site data now initiates the tag section as an empty array instead of undefined.

Every site has an id now. This should hopefully provide a nice key.