Updated tag inclusion flow, styles
Edit 2 button renamed to apply, edit button renamed to edit tags.
Form will close after applying.

When you filter out tags, instead of getting an HTTP update, the site will automatically filter to the applicable questions out of the 25 or so in the panel. It uses styles to hide / show, so that if you remove tags, it will show everything back again (instead of being destructive). This is still not great though, going to improve flow soon.

Updated styles.
Gotta get a good font in there soon.